+++ x/core/02_post_dispatch/client/cli/tx_igp.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
+	"strings"
 
 	"github.com/bcp-innovations/hyperlane-cosmos/util"
 
@@ -102,9 +103,30 @@ func CmdCreateIgp() *cobra.Command {
 
 func CmdSetIgpOwner() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "set-owner [igp-id] [new-owner]",
+		Use:   "set-owner [igp-id]",
 		Short: "Update a Hyperlane Interchain Gas Paymaster",
-		Args:  cobra.ExactArgs(2),
+		Args:  cobra.ExactArgs(1),
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			yes, err := cmd.Flags().GetBool("yes")
+			if err != nil {
+				return err
+			}
+
+			if renounceOwnership && !yes {
+				fmt.Print("Are you sure you want to renounce ownership? This action is irreversible. (yes/no): ")
+				var response string
+
+				_, err := fmt.Scanln(&response)
+				if err != nil {
+					return err
+				}
+
+				if strings.ToLower(response) != "yes" {
+					return fmt.Errorf("canceled transaction")
+				}
+			}
+			return nil
+		},
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
@@ -117,9 +139,10 @@ func CmdSetIgpOwner() *cobra.Command {
 			}
 
 			msg := types.MsgSetIgpOwner{
-				Owner:    clientCtx.GetFromAddress().String(),
-				IgpId:    igpId,
-				NewOwner: args[1],
+				Owner:             clientCtx.GetFromAddress().String(),
+				IgpId:             igpId,
+				NewOwner:          newOwner,
+				RenounceOwnership: renounceOwnership,
 			}
 
 			_, err = sdk.AccAddressFromBech32(msg.Owner)
@@ -130,6 +153,9 @@ func CmdSetIgpOwner() *cobra.Command {
 			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), &msg)
 		},
 	}
+
+	cmd.Flags().StringVar(&newOwner, "new-owner", "", "set updated owner")
+	cmd.Flags().BoolVar(&renounceOwnership, "renounce-ownership", false, "renounce ownership")
 
 	flags.AddTxFlagsToCmd(cmd)