+++ x/core/keeper/msg_mailbox_test.go
@@ -40,7 +40,9 @@ TEST CASES - msg_mailbox.go
 * ProcessMessage (invalid) with invalid metadata (Noop ISM)
 * ProcessMessage (valid) (Noop ISM)
 * ProcessMessage (valid) (Multisig ISM)
-* SetMailbox (valid) with hooks
+* SetMailbox (invalid) with invalid new owner
+* SetMailbox (invalid) with non-owner address
+* SetMailbox (valid) renounce ownership
 * SetMailbox (valid) without hooks
 * SetMailbox (valid)
 
@@ -398,7 +400,7 @@ var _ = Describe("msg_mailbox.go", Ordered, func() {
 		Expect(mailboxId.String()).To(Equal(mailboxId.String()))
 	})
 
-	It("SetMailbox (valid) with hooks", func() {
+	It("SetMailbox (invalid) with invalid new owner", func() {
 		// Arrange
 		mailboxId, requiredHook, defaultHook, ism := createValidMailbox(s, creator.Address, "noop", 1)
 
@@ -409,12 +411,44 @@ var _ = Describe("msg_mailbox.go", Ordered, func() {
 
 		// Act
 		_, err := s.RunTx(&types.MsgSetMailbox{
-			Owner:        sender.Address,
-			MailboxId:    mailboxId,
-			DefaultIsm:   &noopIsmId,
-			DefaultHook:  &defaultHookId,
-			RequiredHook: &requiredHookId,
-			NewOwner:     newOwner,
+			Owner:             creator.Address,
+			MailboxId:         mailboxId,
+			DefaultIsm:        &noopIsmId,
+			DefaultHook:       &defaultHookId,
+			RequiredHook:      &requiredHookId,
+			NewOwner:          newOwner,
+			RenounceOwnership: false,
+		})
+
+		// Assert
+		Expect(err.Error()).To(Equal("invalid new owner"))
+
+		mailbox, err := s.App().HyperlaneKeeper.Mailboxes.Get(s.Ctx(), mailboxId.GetInternalId())
+		Expect(err).To(BeNil())
+		Expect(mailbox.DefaultIsm).To(Equal(ism))
+		Expect(mailbox.DefaultHook).To(Equal(&defaultHook))
+		Expect(mailbox.RequiredHook).To(Equal(&requiredHook))
+		Expect(mailbox.Owner).To(Equal(creator.Address))
+	})
+
+	It("SetMailbox (invalid) with non-owner address", func() {
+		// Arrange
+		mailboxId, requiredHook, defaultHook, ism := createValidMailbox(s, creator.Address, "noop", 1)
+
+		noopIsmId := createNoopIsm(s, sender.Address)
+		defaultHookId := createIgp(s, creator.Address)
+		requiredHookId := createIgp(s, creator.Address)
+		newOwner := i.GenerateTestValidatorAddress("new_owner").AccAddress.String()
+
+		// Act
+		_, err := s.RunTx(&types.MsgSetMailbox{
+			Owner:             sender.Address,
+			MailboxId:         mailboxId,
+			DefaultIsm:        &noopIsmId,
+			DefaultHook:       &defaultHookId,
+			RequiredHook:      &requiredHookId,
+			NewOwner:          newOwner,
+			RenounceOwnership: false,
 		})
 
 		// Assert
@@ -428,21 +462,78 @@ var _ = Describe("msg_mailbox.go", Ordered, func() {
 		Expect(mailbox.Owner).To(Equal(creator.Address))
 	})
 
+	It("SetMailbox (invalid) renounce ownership when new owner is specified", func() {
+		// Arrange
+		mailboxId, requiredHook, defaultHook, ismId := createValidMailbox(s, creator.Address, "noop", 1)
+
+		noopIsmId := createNoopIsm(s, sender.Address)
+
+		// Act
+		_, err := s.RunTx(&types.MsgSetMailbox{
+			Owner:             creator.Address,
+			MailboxId:         mailboxId,
+			DefaultIsm:        &noopIsmId,
+			DefaultHook:       nil,
+			RequiredHook:      nil,
+			NewOwner:          "test",
+			RenounceOwnership: true,
+		})
+
+		// Assert
+		Expect(err.Error()).To(Equal("cannot set new owner and renounce ownership at the same time"))
+
+		mailbox, err := s.App().HyperlaneKeeper.Mailboxes.Get(s.Ctx(), mailboxId.GetInternalId())
+		Expect(err).To(BeNil())
+		Expect(mailbox.DefaultIsm).To(Equal(ismId))
+		Expect(mailbox.DefaultHook).To(Equal(&defaultHook))
+		Expect(mailbox.RequiredHook).To(Equal(&requiredHook))
+		Expect(mailbox.Owner).To(Equal(creator.Address))
+	})
+
+	It("SetMailbox (valid) renounce ownership", func() {
+		// Arrange
+		mailboxId, requiredHook, defaultHook, _ := createValidMailbox(s, creator.Address, "noop", 1)
+
+		noopIsmId := createNoopIsm(s, sender.Address)
+
+		// Act
+		_, err := s.RunTx(&types.MsgSetMailbox{
+			Owner:             creator.Address,
+			MailboxId:         mailboxId,
+			DefaultIsm:        &noopIsmId,
+			DefaultHook:       nil,
+			RequiredHook:      nil,
+			NewOwner:          "",
+			RenounceOwnership: true,
+		})
+
+		// Assert
+		Expect(err).NotTo(HaveOccurred())
+
+		mailbox, err := s.App().HyperlaneKeeper.Mailboxes.Get(s.Ctx(), mailboxId.GetInternalId())
+		Expect(err).To(BeNil())
+		Expect(mailbox.DefaultIsm).To(Equal(noopIsmId))
+		Expect(mailbox.DefaultHook).To(Equal(&defaultHook))
+		Expect(mailbox.RequiredHook).To(Equal(&requiredHook))
+		Expect(mailbox.Owner).To(Equal(""))
+	})
+
 	It("SetMailbox (valid) without hooks", func() {
 		// Arrange
 		mailboxId, requiredHook, defaultHook, _ := createValidMailbox(s, creator.Address, "noop", 1)
 
 		noopIsmId := createNoopIsm(s, sender.Address)
-		newOwner := "new_owner"
+		newOwner := i.GenerateTestValidatorAddress("new_owner").AccAddress.String()
 
 		// Act
 		_, err := s.RunTx(&types.MsgSetMailbox{
-			Owner:        creator.Address,
-			MailboxId:    mailboxId,
-			DefaultIsm:   &noopIsmId,
-			DefaultHook:  nil,
-			RequiredHook: nil,
-			NewOwner:     newOwner,
+			Owner:             creator.Address,
+			MailboxId:         mailboxId,
+			DefaultIsm:        &noopIsmId,
+			DefaultHook:       nil,
+			RequiredHook:      nil,
+			NewOwner:          newOwner,
+			RenounceOwnership: false,
 		})
 
 		// Assert
@@ -463,16 +554,17 @@ var _ = Describe("msg_mailbox.go", Ordered, func() {
 		noopIsmId := createNoopIsm(s, sender.Address)
 		defaultHookId := createIgp(s, creator.Address)
 		requiredHookId := createIgp(s, creator.Address)
-		newOwner := "new_owner"
+		newOwner := i.GenerateTestValidatorAddress("new_owner").AccAddress.String()
 
 		// Act
 		_, err := s.RunTx(&types.MsgSetMailbox{
-			Owner:        creator.Address,
-			MailboxId:    mailboxId,
-			DefaultIsm:   &noopIsmId,
-			DefaultHook:  &defaultHookId,
-			RequiredHook: &requiredHookId,
-			NewOwner:     newOwner,
+			Owner:             creator.Address,
+			MailboxId:         mailboxId,
+			DefaultIsm:        &noopIsmId,
+			DefaultHook:       &defaultHookId,
+			RequiredHook:      &requiredHookId,
+			NewOwner:          newOwner,
+			RenounceOwnership: false,
 		})
 
 		// Assert
@@ -574,7 +666,7 @@ func createNoopIsm(s *i.KeeperTestSuite, creator string) util.HexAddress {
 	return response.Id
 }
 
-func setDestinationGasConfig(s *i.KeeperTestSuite, creator string, igpId util.HexAddress, domain uint32) error {
+func setDestinationGasConfig(s *i.KeeperTestSuite, creator string, igpId util.HexAddress, _ uint32) error {
 	_, err := s.RunTx(&pdTypes.MsgSetDestinationGasConfig{
 		Owner: creator,
 		IgpId: igpId,