+++ CHANGELOG.md
@@ -18,6 +18,7 @@ An '!' indicates a state machine breaking change.
 
 ### Features
 
+- ! [#98](https://github.com/bcp-innovations/hyperlane-cosmos/pull/98) Add renounce ownership functionalities and validate new owner
 - ! [#100](https://github.com/bcp-innovations/hyperlane-cosmos/pull/100) Routing ISM
 
 ### Improvements